pyproject.toml
[project]
name = "fastapi-start-template"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.11"
dependencies = [
    "fastapi[standard]>=0.116.0",
    "loguru>=0.7.3",
    "orjson>=3.10.18",
    "pydantic-settings>=2.10.1",
    "python-dotenv>=1.1.1",
    "slowapi>=0.1.9",
]

[dependency-groups]
dev = [
    "pytest>=8.4.1",
    "ruff>=0.12.2",
]
